refactor: reorganize plugin codebase
