--- src/constants.ts
@@ -1,6 +1,0 @@
-export enum SupportedDiagrams {
-    Mermaid = '.mermaid',
-    Mehrmaid = '.block-language-mehrmaid',
-    PlantUML = '.block-language-plantuml',
-    Graphviz = '.block-language-dot',
-}
