+++ src/typing/typing.ts
@@ -10,3 +10,10 @@ export interface Data {
         nativeTouchEventsEnabled: boolean;
     };
 }
+
+export enum SupportedDiagrams {
+    Mermaid = '.mermaid',
+    Mehrmaid = '.block-language-mehrmaid',
+    PlantUML = '.block-language-plantuml',
+    Graphviz = '.block-language-dot',
+}
